requirements.txt
boto3==1.17.86
botocore==1.20.86
certifi==2021.5.30
chardet==4.0.0
idna==2.10
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
